Update logrotate configuration to retain log file.
Updated logrotate(8) configuration to retain the original log file
but copy its content to a new rotated file and truncate the old file.

This fixes bug BZ#917560. The glitch was, logrotate(8) would rename
an old log file. But the server would continue to append logs to the
same rotated file, instead of creating the default one, because it
already has an open file descriptor of the old file.

The `copytruncate' directive tells logrotate(8) to copy an old
file, instead of renaming it.

  -> https://bugzilla.redhat.com/show_bug.cgi?id=917560