Fix an off-by-one error.
Changed server.c to not over step the end-of-string marker while
reading comma separated list of IP addresses from a configuration
file.

The issue was that, when the list of IP addresses did not end with
a comma(','), server would crash by trying to read an IP address
from invalid memory area.