Enable Systemd(1) & Xinetd(8) support for axfrdns
Added Systemd(1) & Xinetd(8) configurations to support invocation
of the axfrdns(8) zone transfer server. It used to depend on a
xinetd(8) like service called 'tcpserver' for creation of a network
socket. Now the same is done via Systemd(1) & Xinetd(8) services.

Also changed axfrdns server to redirect debug messages to a log file
when running as daemon.