Accept flag parameter in redirect_to_log
Earlier 'redirect_to_log' routine used to direct both 'stdout'
and 'stderr' of a server to its log file. But axfrdns(8) server
requires only 'stderr' to be redirected to its log file. Because
'stdout' is redirected to a network socket by Xinetd(8) or
Systemd(1) service.

Changed 'redirect_to_log' routine to accept a flag parameter.
It'll decide which standard descriptors to redirect to a log file.
It could be an 'OR' of 'STDOUT_FILENO' & 'STDERR_FILENO'.

Changed the server sources to pass this new parameter while calling
'redirect_to_log' routine.