Update axfrdns' xinetd & systemd configurations
axfrdns(8) server depends on Xinetd(8) or Systemd(1) like services
to create a network socket and start the daemon process. axfrdns(8)
need not fork a sub-process to start this daemon.

Changed axfrdns(8) server to avoid creation of a daemon process;
And to only direct its standard error(STDERR_FILENO) descriptor to
a log file. As standard output is directed to a network socket by
its parent process.

Also tweaked the xinetd(8) & Systemd(1) configurations to toggle
the 'wait' parameter and to pass '-D' server argument.

Thank you Edwin Eefting(github.com/psy0rz) for a comprehensive
patch review and for helping to fix the axfrdns(8) server. :)